Fix FeedItem.generate_thumbnail
There are two parts to this fix:

The built-in ruby open() does not support redirects properly, so we
installed the httparty gem.

Additionally, the ruby-readability gem doesn't automatically convert
absolute URIs to URLs. We added a check to prepend the base URL to the
URI for images with absolute URIs before processing the image.

Something to consider: currently, the generate_thumbnail only pulls
images that are inside of <a> tags, e.g., <a><img></a>. We may want to
fix this so that it considers every image.